Added functionality for computer move
Function computer_move uses simple logic to decide the computers next
move.  If the computer can when, it takes the winning move.  If not, and
the player can when on the next move, the computer blocks.  Otherwise,
the computer chooses the next move from BEST_MOVES, which places
priority on certain spaces (center, corners, and sides, in that order).
Fixed indentation error in winner function.